workspace/7_framework/fastapi/day06/requirements.txt
fastapi>=0.115.0
pydantic>=1.10.0
pyjwt>=2.4.0
uvicorn>=0.30.0
python-multipart>=0.0.20
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
httpx>=0.24.0
flake8>=5.0.0
